requirements_performance.txt
# Redis Configuration
redis==5.0.1
redis-py-cluster==2.1.3

# Async Processing
aiohttp==3.9.1
aiofiles==23.2.0
asyncio-throttle==1.0.2

# Database Performance
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
sqlalchemy-utils==0.41.1

# Background Jobs
celery==5.3.4
flower==2.0.1

# Monitoring & Metrics
psutil==5.9.6
prometheus-client==0.19.0
structlog==23.2.0

# Compression & Optimization
python-multipart==0.0.6
Pillow==10.1.0
PyPDF2==3.0.1

# Cloud Storage
boto3==1.34.0
google-cloud-storage==2.10.0
azure-storage-blob==12.19.0

# Load Balancing & Networking
gunicorn==21.2.0
gevent==23.9.1
requests-toolbelt==1.0.0

# Security & Rate Limiting
flask-limiter==3.5.0
flask-compress==1.14

# Development & Testing
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
memory-profiler==0.61.0
